Dice rolly,py
This is a beginner level program to roll "dice_count" dices as input from the users to either roll a "target_sum" given from the user or just roll all the dices and give the probability of that sum to happen. I have tried to include all possible errors including if the input of the user is yes/no. The only problem you could run into is if the dice_count number is higher than 375 as it would give the error "OverflowError: int too large to convert to float".