V2 of Dice rolling.py
This updated version now handles up to 10000 rolls with no memory issues and is extremely fast up to 1000.  Everything has been modified to support the usage of dices with a different number of faces than 6 but a function to ask the user for the input on choosing the dice type has not been yet implemented. Also fixed a bug where the program would ask the user to "Roll Again?" and still roll after the user typed in "No".